Factor out metafunctions from Algorithm.hpp
- We want to be able to reuse the metafunctions in the mocking framework without
  having to include all of Algorithm.hpp
